initial refactor to using uv
