+++ pyproject.toml
@@ -1,0 +1,24 @@
+[project]
+name = "web-envs"
+version = "0.1.0"
+description = "Web Environments Task Recorder"
+authors = [{ name = "Joan Cabezas", email = "joan.santiago.cabezas@gmail.com" }]
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "browser-use>=0.7.10",
+    "dspy>=3.0.3",
+    "litellm>=1.77.5",
+    "playwright>=1.55.0",
+    "python-dotenv>=1.1.1",
+]
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+package-dir = { "" = "src" }
+
+[tool.ruff.lint]
+ignore = ["F722"]
