+++ src/tasks/db_to_jsonl_format.py
@@ -6,7 +6,6 @@ from dataclasses import dataclass
 from typing import List, Dict, Any, Optional
 from pathlib import Path
 from html.parser import HTMLParser
-
 from config.storage_config import DATA_DIR
