+++ src/utils/get_tasks_video_path.py
@@ -1,8 +1,8 @@
 import os
 
-from src.utils.get_iso_datetime import get_iso_datetime
-from src.config.storage_config import VIDEOS_DIR
-from src.tasks.task import TaskManager
+from utils.get_iso_datetime import get_iso_datetime
+from config.storage_config import VIDEOS_DIR
+from tasks.task import TaskManager
 import logging
 
 logger = logging.getLogger(__name__)
